fondum/fondum_requirements.txt
appdirs==1.4.3
babelfish==0.5.5
Babel==2.5.0
bcrypt==3.1.3
boto==2.46.1
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
click==6.7
Flask==0.12.1
Flask-Bcrypt==0.7.1
Flask-Login==0.4.0
flask-mongoengine==0.9.3
Flask-OAuthlib==0.9.4
Flask-QRcode==0.10.0
Flask-WTF==0.14.2
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
mongoengine==0.13.0
oauthlib==2.0.2
olefile==0.44
packaging==16.8
Pillow==4.1.1
PLOD==0.1.7
pycparser==2.17
pymongo==3.4.0
pyparsing==2.2.0
python-creole==1.3.1
qrcode==5.3
requests==2.18.4
requests-oauthlib==0.8.0
sendgrid==5.3.0
six==1.10.0
stripe==1.75.3
urllib3==1.22
Werkzeug==0.12.1
WTForms==2.1
